oscaroutput: fix gcc 5.1 warning
An enum was printed as an integer. Casting it before fixes the warnings.

Refs #1644.

+++ src/oscaroutput.cc
@@ -206,7 +206,7 @@ void OscarOutput<Format, Contents>::at_interaction(
           file_.get(),
           "# interaction in %zu out %zu rho %12.7f weight %12.7f type %5i \n",
           incoming_particles.size(), outgoing_particles.size(), density,
-          total_cross_section, process_type);
+          total_cross_section, static_cast<int>(process_type));
     } else {
       /* OSCAR line prefix : initial final
        * particle creation: 0 1
@@ -217,7 +217,7 @@ void OscarOutput<Format, Contents>::at_interaction(
        */
       std::fprintf(file_.get(), "%zu %zu %12.7f %12.7f %5i \n",
                    incoming_particles.size(), outgoing_particles.size(),
-                   density, total_cross_section, process_type);
+                   density, total_cross_section, static_cast<int>(process_type));
     }
     for (const auto &p : incoming_particles) {
       write_particledata(p);
